fix channel handling in busy case
previously, on a lantiq_requester() call we would blindly create a new
channel object, even if that channel already had one. This caused
playing of the busy sound in the running phone call.

now, we check whether there's already an existing channel object (which
means the channel is busy) and exit early.